+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "input-masking",
-  "version": "1.0.4",
+  "version": "0.0.1",
   "description": "Styleable, Accessble Input Masking",
-  "main": "index.js",
+  "main": "dist/masked-input.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "storybook": "start-storybook -p 9001",
@@ -24,8 +24,7 @@
   },
   "homepage": "https://github.com/stephencweiss/input-masking#readme",
   "dependencies": {
-    "react": "^16.10.1",
-    "styled-components": "^4.4.0"
+    "react": "^16.10.1"
   },
   "devDependencies": {
     "@babel/core": "^7.6.2",
